feat+refactor(engine): Reogranize auth module + add context dict target for AuthSandbox
