+++ tracecat/cli/workflow.py
@@ -6,7 +6,7 @@ import orjson
 import rich
 import typer
 
-from tracecat.auth import AuthenticatedAPIClient
+from tracecat.auth.clients import AuthenticatedAPIClient
 from tracecat.dsl.workflow import DSLInput
 from tracecat.types.api import WebhookResponse
