+++ tracecat/contexts.py
@@ -7,7 +7,7 @@ from typing import TYPE_CHECKING
 from pydantic import BaseModel
 
 if TYPE_CHECKING:
-    from tracecat.auth import Role
+    from tracecat.auth.credentials import Role
 
 
 class RunContext(BaseModel):
